refactor(fast_io): decompose splice.rs into submodules (SPL-18 #2445)
